requirements.txt
fastapi
uvicorn
tensorflow
numpy
python-multipart
pillow
opencv-python
pydantic
python-dotenv
jinja2
aiofiles